Ban Generic Lists
This commit updates several IXList interfaces and implementations to
ban generic Lists and replace them with arrays + array size fields. The
reason for this is that a lot of time is spent getting and upcasting
values from Generic Collections.
Currently, this change accounted for ca. 10% performance boost on
Library Loading and 40% performance boost on Compilation.