Improved If Expressions
Previously, when compiling an if Statement in the place of an
Expression, it had to invoke MethodWriter.pop after the then block to
prepare the stack for the else block, so it does not end up with two
values on the stack. This behavior has now been moved to the
MethodWriterImpl.