Improved Local Variables
Local Variable Declaration are not parsed by the ExpressionParser
anymore, but instead by the more powerful StatementListParser that
utilizes an underlying TypeParser. A side of this is that local
variables cannot be used as Expressions any more, but since that did
not make sense in the first place, this should not be a problem.