Improved Array Syntax
Arrays are now denoted with a square bracket syntax like [ 1, 2, 3 ].
This change helps to separate arrays from statement lists in order to
avoid ambiguities.
Another improvement is that the ExpressionParser does not use an
underlying TypeParser for the first Identifier that could have been a
static access. This has been replaced with a simple FieldAccess that
can now also resolve to a ClassAccess if it can't be resolved to a
field, method or type.