nil
Added the nil literal / keyword as well as the NilConvertible
interface. The nil literal is used to describe an empty value, but not
null. It can be used in the place of arrays (where it aliases an empty
array [ ]) or in the place of any type that implements the
NilConvertible marker interface.

Other changes in this commit:
- Fixed generation of the implicit no-arg constructor of case classes
- Improved the error messages of invalid literal conversions
- Minor improvements and bugfixes