Type Inference
Added support for type inference via the 'var' keyword. Types are
inferred after RESOLVE and before CHECK_TYPES using local type
inference. Each member has their own custom errors when a type cannot
be inferred. In addition to type inference, the '...' special symbol
has been added that functions as a wildcard value. This value with be
the default value of the type of it's context, i.e. false, 0 or null.
Closes #25