Variable Assignment Capture
Variables that are assigned from withing a Lambda Expression will now
be converted to a 'Ref Variable' that holds an object containing the
state of the variable. These objects are instances of the
dyvil.lang.ref.*Ref classes, and all accesses and assignments to the
variable will be redirected to the Reference Object.