Prefix and Postfix Operator Precedence
Prefix and Postfix operators can now have precedence against other
(infix) operators. By default, they have a precedence of 1000, which
means they are always executed first. However, it is also possible to
define a different one using the usual syntax. Setting a precedence of
1 for the 'println' "operator" will always execute it last, which means
that it can easily handle expressions like the following:

println "The value is: " + value

Other changes in this commit:
- Amplified Standard Operator precedence values.
- Improved ExpressionParser implementation.