Remove - as part of number literals
A minus sign - at the beginning of a number literal is no longer
handled by the Dyvil Lexer. The parser implementation has to provide
the semantics needed for negating number literals.

The implementations of the DPF Parser and Pattern Parsers have been
updated. The default Expression Parser supports this without changes.