Update Primitive Instance Intrinsics
Converted all Primitive Instance methods with Intrinsic Annotations to
static infix methods. This allows us to make several simplifications in
the compiler and avoid errors that might come up in the future because
of boxing or the lack thereof. The update also includes some minor
tweaks to the Primitive Classes.